Split out test implementation from tests themselves
