+++ src/remote/BUILD
@@ -2,7 +2,10 @@ go_library(
     name = "remote",
     srcs = glob(
         ["*.go"],
-        exclude = ["*_test.go"],
+        exclude = [
+            "*_test.go",
+            "test_impl.go",
+        ],
     ),
     visibility = ["PUBLIC"],
     deps = [
@@ -23,7 +26,10 @@ go_library(
 
 go_test(
     name = "remote_test",
-    srcs = ["remote_test.go"],
+    srcs = [
+        "remote_test.go",
+        "test_impl.go",
+    ],
     data = ["test_data"],
     deps = [
         ":remote",
@@ -33,6 +39,5 @@ go_test(
         "//third_party/go:protobuf",
         "//third_party/go:remote-apis",
         "//third_party/go:rpcstatus",
-        "//third_party/go:testify",
-    ],
+        "//third_party/go:testify",    ],
 )
